Tweak migrator
