+++ tools/TestErrorInstanceMigration/AppHost/AppHost.cs
@@ -1,9 +1,9 @@
 using AppHost;
 using Particular.Aspire.Hosting.ServicePlatform.Platform;
 
-var mode = MigrationMode.PostMigrationMode;
-var enableIngestion = true;
+var mode = MigrationMode.Step1PreMigration;
 var targetPersistence = TargetPersistence.SqlServer;
+var enableIngestion = mode != MigrationMode.Step2RetryMessages;
 var imageTag = "latest";
 
 Console.WriteLine($"Migration mode: {mode}; target persistence: {targetPersistence}; ServiceControl image tag: {imageTag}");
@@ -26,14 +26,20 @@ var targetDatabase = builder.AddTargetDatabase(targetPersistence);
 var errorInstance =
     mode switch
     {
-        MigrationMode.PreMigration => platform.AddRavenErrorInstance(raven, imageTag, enableIngestion),
-        MigrationMode.PostMigrationMode => platform.AddSqlErrorInstance(raven, targetPersistence, targetDatabase, imageTag),
+        MigrationMode.Step1PreMigration or MigrationMode.Step2RetryMessages => platform.AddRavenErrorInstance(raven, imageTag, enableIngestion),
+        MigrationMode.Step3PostMigration => platform.AddSqlErrorInstance(raven, targetPersistence, targetDatabase, imageTag),
         _ => throw new Exception()
     };
 
 platform.AddServicePulse("servicepulse", errorInstance!);
 
 builder.AddProject<Projects.FailingEndpoint>("failing-endpoint")
+    .WithEnvironment("CREATE_FAILURES", (mode == MigrationMode.PreMigration).ToString())
+    .WithUrlForEndpoint("http", url =>
+    {
+        url.DisplayText = "Create Errors";
+        url.Url += "/createerrors";
+    })
     .WithParticularPlatform(platform);
 
 await builder.Build().RunAsync();
