+++ tools/TestErrorInstanceMigration/FailingEndpoint/FailDeterministically.cs
@@ -1,0 +1,5 @@
+public sealed class FailDeterministically : IMessage
+{
+    public Guid ErrorId { get; set; }
+    public string RandomPayload { get; set; } = string.Empty;
+}
