+++ tools/TestErrorInstanceMigration/FailingEndpoint/FailingEndpoint.csproj
@@ -13,6 +13,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ReplacingErrorInstance.ServiceDefaults\ReplacingErrorInstance.ServiceDefaults.csproj" />
+    <ProjectReference Include="..\ServiceDefaults\ServiceDefaults.csproj" />
   </ItemGroup>
 </Project>
