+++ tools/TestErrorInstanceMigration/FailingEndpoint/SimulatedDeterministicFailure.cs
@@ -1,0 +1,1 @@
+public sealed class SimulatedDeterministicFailure(string message) : Exception(message);
