+++ tools/TestErrorInstanceMigration/ReplacingErrorInstance.slnx
@@ -1,5 +1,5 @@
 <Solution>
   <Project Path="AppHost\AppHost.csproj" />
   <Project Path="FailingEndpoint/FailingEndpoint.csproj" />
-  <Project Path="ReplacingErrorInstance.ServiceDefaults/ReplacingErrorInstance.ServiceDefaults.csproj" />
+  <Project Path="ServiceDefaults\ServiceDefaults.csproj" />
 </Solution>
