+++ tools/TestErrorInstanceMigration/ServiceDefaults/ParticularPlatformConfig.cs
