+++ tools/TestErrorInstanceMigration/ServiceDefaults/ServiceDefaults.csproj
@@ -5,6 +5,7 @@
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
         <IsAspireSharedProject>true</IsAspireSharedProject>
+        <RootNamespace>ReplacingErrorInstance.ServiceDefaults</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
